+++ src/compiler/serialization/mod.ts
@@ -1,6 +1,4 @@
 export * from './ucs-compiler.js';
-export * from './ucs-enable-plain-text.js';
-export * from './ucs-enable-uri-encoded.js';
 export * from './ucs-export.signature.js';
 export * from './ucs-formatter.js';
 export * from './ucs-function.js';
@@ -17,4 +15,6 @@ export * from './ucs-process-primitives.js';
 export * from './ucs-process-string.js';
 export * from './ucs-process-unknown.js';
 export * from './ucs-setup.js';
+export * from './ucs-support-plain-text.js';
+export * from './ucs-support-uri-encoded.js';
 export * from './ucs-writer.class.js';
